Move type information files to ./src/type folder
